Fixing with double VPFloat

+++ include/DecisionNode.hpp
@@ -8,7 +8,7 @@
 namespace Kaadugal
 {
     // Use Variable Precision Floating point representation
-    typedef float VPFloat; // Change to float or long double as needed
+    typedef double VPFloat; // Change to float or long double as needed
     enum NodeType
     {
 	SplitNode,
